requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-mistralai>=0.1.0
mistralai>=0.1.0
chromadb>=0.4.22
python-dotenv>=1.0.0
click>=8.1.7
tiktoken>=0.5.2
python-multipart>=0.0.6